Aula automacao/Controle_PC/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
ddgs
requests
python-dotenv
pycaw
comtypes
screen_brightness_control